break out a few functions
